+++ src/cache.c
@@ -24,22 +24,22 @@ typedef struct FontCache {
 ALLEGRO_BITMAP *GetNewlyAddedBitmapFromCache(char *asset_path);
 
 static BitmapCache *bitmap_cache;
-static unsigned int bitmap_cache_size  = 256;
-static unsigned int bitmap_cache_index = 0;
+static unsigned short int bitmap_cache_size  = 256;
+static unsigned short int bitmap_cache_index = 0;
 
 static FontCache *font_cache;
-static unsigned int font_cache_size  = 256;
-static unsigned int font_cache_index = 0;
+static unsigned short int font_cache_size  = 256;
+static unsigned short int font_cache_index = 0;
 
 void InitializeCache() 
 {
 
     bitmap_cache = malloc(sizeof(BitmapCache) * bitmap_cache_size);
-    for (unsigned int i = 0; i < bitmap_cache_size; i++)
+    for (unsigned short int i = 0; i < bitmap_cache_size; i++)
         bitmap_cache[i].asset_path = NULL;
 
     font_cache = malloc(sizeof(FontCache) * font_cache_size);
-    for (unsigned int i = 0; i < font_cache_size;i++)
+    for (unsigned short int i = 0; i < font_cache_size;i++)
         font_cache[i].asset_path = NULL;
 
 }
@@ -50,7 +50,7 @@ ALLEGRO_BITMAP *GetBitmapFromCache(char *asset_path)
     if (asset_path == NULL || strcmp(asset_path,"") == 0)
         return NULL;
 
-    for (unsigned int i = 0; i < bitmap_cache_index; i++) {
+    for (unsigned short int i = 0; i < bitmap_cache_index; i++) {
 
         if (strcmp(asset_path, bitmap_cache[i].asset_path) == 0)
             return bitmap_cache[i].bitmap;
@@ -61,7 +61,7 @@ ALLEGRO_BITMAP *GetBitmapFromCache(char *asset_path)
 
 }
 
-void IncreaseCacheSize(unsigned int cache_index, unsigned int *cache_size, int size_of_cache_struct, void *cache)
+void IncreaseCacheSize(unsigned short int cache_index, unsigned short int *cache_size, int size_of_cache_struct, void *cache)
 {
 
     if (cache_index + 1 < *cache_size)
@@ -122,7 +122,7 @@ ALLEGRO_FONT *GetFontFromCache(char *asset_path, int font_size)
     if (asset_path == NULL || font_size == 0)
         return NULL;
 
-    for (size_t i = 0; i < font_cache_index; i++) {
+    for (unsigned short int i = 0; i < font_cache_index; i++) {
 
         if (font_cache[i].font_size == font_size && strcmp(font_cache[i].asset_path, asset_path) == 0) 
             return font_cache[i].font;
